// src/api/generated/model/jobDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GearFlow.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JobDtoPriceCents } from './jobDtoPriceCents';

export interface JobDto {
  id: string;
  name: string;
  description: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  priceCents: JobDtoPriceCents;
}